Rename project to "seal"
"shield" was something of a misnomer because this project can't actually
protect your files from being corrupted. All it can do is alert you that
a file was corrupted so you can take action.

In general, think of safety seals you find on medicine bottles and jugs
of milk. If it's broken, you should tread carefully.